Speak with current text document service
