+++ src/main/java/io/rathr/audrey/lsp/AudreyServer.java
@@ -1,27 +1,72 @@
 package io.rathr.audrey.lsp;
 
+import org.eclipse.lsp4j.DidChangeConfigurationParams;
 import org.eclipse.lsp4j.DidChangeTextDocumentParams;
+import org.eclipse.lsp4j.DidChangeWatchedFilesParams;
 import org.eclipse.lsp4j.DidCloseTextDocumentParams;
 import org.eclipse.lsp4j.DidOpenTextDocumentParams;
 import org.eclipse.lsp4j.DidSaveTextDocumentParams;
-import org.eclipse.lsp4j.Hover;
 import org.eclipse.lsp4j.InitializeParams;
 import org.eclipse.lsp4j.InitializeResult;
 import org.eclipse.lsp4j.ServerCapabilities;
-import org.eclipse.lsp4j.TextDocumentPositionParams;
 import org.eclipse.lsp4j.TextDocumentSyncKind;
+import org.eclipse.lsp4j.jsonrpc.Launcher;
+import org.eclipse.lsp4j.launch.LSPLauncher;
 import org.eclipse.lsp4j.services.LanguageClient;
 import org.eclipse.lsp4j.services.LanguageClientAware;
 import org.eclipse.lsp4j.services.LanguageServer;
 import org.eclipse.lsp4j.services.TextDocumentService;
 import org.eclipse.lsp4j.services.WorkspaceService;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.net.ServerSocket;
+import java.net.Socket;
 import java.util.concurrent.CompletableFuture;
+import java.util.logging.Logger;
 
-public class AudreyServer implements LanguageServer, TextDocumentService, LanguageClientAware {
+public class AudreyServer implements LanguageServer, LanguageClientAware {
+    static {
+        System.setProperty(
+            "java.util.logging.SimpleFormatter.format",
+            "[%1$tF %1$tT] [AUDREY] [%4$-7s] %5$s %n"
+        );
+    }
+
+    static final Logger LOG = Logger.getLogger(AudreyServer.class.getName());
+
+
+    final static int DEFAULT_PORT = 8123;
+
     private final TextDocumentService textDocumentService;
     private String workspaceRoot;
     private LanguageClient client;
+
+    public static void main(String[] args) {
+        try {
+            LOG.info("Listening on port " + DEFAULT_PORT);
+            final ServerSocket serverSocket = new ServerSocket(DEFAULT_PORT);
+
+            LOG.info("Waiting on client connection");
+            final Socket socket = serverSocket.accept();
+
+            InputStream in = socket.getInputStream();
+            OutputStream out = socket.getOutputStream();
+
+            AudreyServer server = new AudreyServer();
+            final Launcher<LanguageClient> launcher = LSPLauncher.createServerLauncher(server, in, out);
+
+            LanguageClient client = launcher.getRemoteProxy();
+            server.connect(client);
+
+            launcher.startListening();
+            LOG.info("Client connected");
+        } catch (IOException e) {
+            LOG.severe(e.getMessage());
+            e.printStackTrace();
+        }
+    }
 
     public AudreyServer() {
         this.textDocumentService = new AudreyTextDocumentService();
@@ -32,7 +77,8 @@ public class AudreyServer implements LanguageServer, TextDocumentService, Langua
         workspaceRoot = params.getRootUri();
 
         final ServerCapabilities capabilities = new ServerCapabilities();
-        capabilities.setTextDocumentSync(TextDocumentSyncKind.None);
+        capabilities.setTextDocumentSync(TextDocumentSyncKind.Full);
+        capabilities.setCodeActionProvider(false);
         capabilities.setHoverProvider(true);
 
         return CompletableFuture.completedFuture(new InitializeResult(capabilities));
@@ -55,32 +101,15 @@ public class AudreyServer implements LanguageServer, TextDocumentService, Langua
 
     @Override
     public WorkspaceService getWorkspaceService() {
-        return null;
-    }
+        return new WorkspaceService() {
+            @Override
+            public void didChangeConfiguration(final DidChangeConfigurationParams params) {
+            }
 
-    @Override
-    public CompletableFuture<Hover> hover(final TextDocumentPositionParams position) {
-        return null;
-    }
-
-    @Override
-    public void didOpen(final DidOpenTextDocumentParams params) {
-
-    }
-
-    @Override
-    public void didChange(final DidChangeTextDocumentParams params) {
-
-    }
-
-    @Override
-    public void didClose(final DidCloseTextDocumentParams params) {
-
-    }
-
-    @Override
-    public void didSave(final DidSaveTextDocumentParams params) {
-
+            @Override
+            public void didChangeWatchedFiles(final DidChangeWatchedFilesParams params) {
+            }
+        };
     }
 
     @Override
